Move websocket specific code to the transport folder.
