+++ main.go
@@ -15,7 +15,7 @@ package main
 import (
 	"flag"
 	jsonRpcApi "github.com/eclipse/che-machine-exec/api/jsonrpc"
-	"github.com/eclipse/che-machine-exec/api/websocket"
+	"github.com/eclipse/che-machine-exec/transport/websocket"
 	"github.com/eclipse/che/agents/go-agents/core/jsonrpc"
 	"github.com/eclipse/che/agents/go-agents/core/jsonrpc/jsonrpcws"
 	"github.com/eclipse/che/agents/go-agents/core/rest"
