+++ transport/websocket/attach.go
